:recycle: [Refactor] 공통 파일 합치기
