--- src/main/java/ssuPlector/dto/requestDto/ProjectListRequestDto.java
@@ -1,16 +1,0 @@
-package ssuPlector.dto.requestDto;
-
-import jakarta.annotation.Nullable;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class ProjectListRequestDto {
-    @Nullable private String searchString;
-    @Nullable private String category;
-    @Nullable private String sortType; // recent, old, high, low
-}
